+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 2.0.0
+--------------------
+- Fix: Editors will now use portal language if user language is null
+- Fix: `author` and `created` fields are now passed correctly
+- Fix: `firstname` and `lastname` fields are deprecated, using `name` instead
+
+- Enh: Better way to pass editors config
+
 ## 1.0.2
 --------------------
 - Fix: PHP 7.2 compatibility issues